Add an all sky NVSS,SUMSS,MGPSII catalogue.
This is in the form of a FITS binary table with 3 columns
RA(2000), DEC(2000), PEAK. The table was constructed by transferring
the peak flux from its measured frequency, 843MHz of MGPSII, SUMSS
and 1,4GHz  of NVSS to a reference frequency of 1284MHz using a spectral
index of -0.7. The reference frequency is contained in the FITS header
under the 'REF FREQ' keyword.

The FITS table is structured so that it can appropriately be read using
the functions in the demerit.py module.